==> R/03-cox_analysis_low.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(skimr)
library(ecotox)
library(survival)
library(coxme)
library(stats)
library(AICcmodavg) # used this for the AICcmodavg function to calculate AICcs quickly
library(janitor)
library(broom)
library(ehahelper) # used this for the tidy function to extract values out of coxme.  https://rdrr.io/github/junkka/ehahelper/
library(survminer)
library(ggplot2)
```

# Data

```{r}
mort_dat <- read_rds(here::here("data", "derived", "mort_dat_final.rds"))
```

Create a new column in data frame for group (e.g., ULM-1, NNP-2)

```{r}
mort_dat <- mort_dat %>%
  mutate(group = str_sub(batch, end = -5)) %>%
  select(tag,group,batch,everything())
```

Create a new column in the data frame that represents "recovery."
* Seems the low salinity treatment ended on 10/20/22 and high salinity treatment ended on 11/01/2022.
* All animals in low treatment with death date on or before 10/20/22 will be considered "n" for recovery, all others in low treatment will get "y."
* All animals in high treatment with death date on or before 11/01/22 will be considered "n" for recovery, all others in low treatment will get "y."

Oysters pulled on day 9 for low salinity (day 1 = 10/12/2022)
Oysters pulled on day 21 for high salinity (day 1 = 10/12/2022)

```{r}
mort_dat <- mort_dat %>%
  mutate(recovery = if_else(death_date >= "2022-10-21" & treatment == "L","y",
                    if_else(death_date >= "2022-11-02" & treatment == "H","y","n")))
```

Add in a censoring variable to data
If oysters survived or died in recovery (recovery = "y"), then they are censored "1", otherwise they are given "2", coding for mortality. 

```{r}
mort_dat <- mort_dat %>%
  mutate(censored = if_else(recovery == "y" | mortality == 0,1,2))
```

# Hypothesis Set

Alternative Hypotheses

Response: Days to Death (numeric)
Unit: Inidividual
r() indiciates random effect

	  	i.  Region + r(Tank) 
			ii. Region + r(Tank) + Site  
			iii.Region + r(Tank) + Site + Length
			iv. Region + r(Tank) + Site + Length + r(Group)
			v.  Region + r(Tank) + Site + Region x Site
			vi. Region + r(Tank) + Site + Region x Site + Length
			vii. Region + r(Tank) + Site + Region X Site + Length + r(Group)


# Data for Cox Models

Only considering data from the start of the test salinities (day "0" and beyond)
There are only a few, highly influential observations that occur before time 0.  Those mortalities are unlikely to be closely associated with the desired challenge.

```{r}
data <- mort_dat %>%
  filter(treatment == "L") %>%
  filter(day_of_death >= 0)
```

```{r}
data$region<-as.factor(data$region)
data$growout_site<-as.factor(data$growout_site)
```

```{r}
data <- data %>%
  mutate(region = fct_relevel(region,c("S","N"))) %>%
  mutate(growout_site = fct_relevel(growout_site,c("CCB","COP")))
```

```{r}
data %>%
  group_by(region) %>%
  summarize(ded = sum(censored == 2),
            tot = n())
```


# Run Cox Models

```{r}
c1 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank), 
                 data = data)
summary(c1)
```

```{r}
c2 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site, 
                 data = data)
summary(c2)
```

```{r}
c3 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site + length_mm,
                 data = data)
summary(c3)
```

```{r}
c3_fixed <- coxph(Surv(day_of_death,censored) ~ region + growout_site + length_mm, data = data)
summary(c3_fixed)

t<-survfit(Surv(day_of_death,censored) ~ region + growout_site + length_mm, data = data)
```


```{r}
c4 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site + length_mm + (1|group/region),
                 data = data)
summary(c4)
```

```{r}
c5 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site + region * growout_site, 
                 data = data)
summary(c5)
```

```{r}
c6 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site + region * growout_site + length_mm, 
                 data = data)
summary(c6)
```

```{r}
c7 <- coxme(Surv(day_of_death,censored) ~ region + (1|tank) + growout_site + region * growout_site + length_mm + (1|group/region),
                 data = data)
summary(c7)
```

# Model Selection and Interpretation

```{r}
models <- list(c1,c2,c3,c4,c5,c6,c7)

low_models_parameters <- tibble(aictab(models))
```

Save data from models for making tables 

```{r}
saveRDS(low_models_parameters, file = here::here("data","derived","low_models_parameters"))
```


AICc calculated from:  AICc = AIC + 2k(k + 1) / (n - k - 1)
where AIC = 2k - 2*Integrated Log Lik

Best models are: 

iii. Region + r(Tank) + Site + Length

vi.  Region + r(Tank) + Site + Region x Site + Length


```{r}
# Extract coefficients and other information from model A
summary_A <- summary(c3)

mort <- summary_A$n[1]
total <- summary_A$n[2]

c3_summary<-tidy(c3)
c3_table <- c3_summary %>%
  mutate(HR = exp(estimate)) %>%
  mutate(HR.conf.low = exp(estimate - 1.96 * std.error)) %>%
  mutate(HR.conf.high = exp(estimate + 1.96 * std.error)) %>%
  select(term, estimate, std.error, conf.low, conf.high, HR, HR.conf.low, HR.conf.high, statistic, p.value)
c3_table

c3_r_effects<-as_tibble(ranef(c3)) %>%
  rownames_to_column() %>%
  rename(tank = rowname, coefficients = tank) %>%
  mutate(hr = exp(coefficients))
```


Save data for tables for best fitting model (m3) for a separate markdown

```{r}
saveRDS(mort, file = here::here("data","derived","04_best_model_low_mort"))

saveRDS(total, file = here::here("data","derived","04_best_model_low_total"))

saveRDS(c3_table, file = here::here("data","derived","04_best_model_low_fixed_effects"))

saveRDS(c3_r_effects, file = here::here("data","derived","04_best_model_low_random_effects"))
```


Interpret model c3: 

* The z values and p values show Region and length have a significant effect on chance of survival. Growout_site with p value just above 0.05

* The coefficient value for regionS represents the hazard associated with the variable regionS.  Because it is negative, it means being in the S region has a negative association with the event (death) compared to the the baseline region, "N"

* The exp(coef), or hazard ratio for southern oysters, is 0.32, meaning the hazard rate in the south is 0.32 times that in the north.

* The coefficient value for growout_siteCOP is negative meaning being in COP increases the time of survival compared to the baseline region, "CCB"

* Being at COP has a hazard ratio of 0.69, meaning the hazard rate at COP is 0.69 times that at CCB

* For each one unit increase in length (mm), there is roughly 3% increase in hazard rate

* Tank effects explain very little of the data. See hazard ratios in c3_r_effects, almost no relative impact on risk of event.

```{r}
summary_B <- summary(c6)

c6_summary<-tidy(c6)
c6_table <- c6_summary %>%
  mutate(HR = exp(estimate)) %>%
  mutate(HR.conf.low = exp(estimate - 1.96 * std.error)) %>%
  mutate(HR.conf.high = exp(estimate + 1.96 * std.error)) %>%
  select(term, estimate, std.error, conf.low, conf.high, HR, HR.conf.low, HR.conf.high, statistic, p.value)

c6_r_effects<-as_tibble(ranef(c6)) %>%
  rownames_to_column() %>%
  rename(tank = rowname, coefficients = tank) %>%
  mutate(hr = exp(coefficients))
```

Interpret model c6: 

* The z values and p values show Region, growout_site, and length have a significant effect on chance of survival.  The interaction term is not significant, thus this model is very similar to model c3. 

* Tanks explained very little of variance.

# Diagnostics

1) Proportional hazards assumption

From Kuitunen et al. 2021: 

"The fundamental assumption in the Cox model is that the hazards are proportional, which means that the relative hazard remains constant over time with different predictor or covariate levels."

Look at the Schoenfel Residuals vs. time to examine this assumption. 

Schoenfeld residuals measure the difference between the observed covariate values and their expected values at each event time for each individual in the dataset. When plotted against time, they indicate the effect of a covariate on the hazard rate over time. 

We will use the fixed version of model three (c3_fixed) to examine residual to test the proportion hazard assumption.

```{r}
test<-cox.zph(c3_fixed)
ggcoxzph(test)
```

The lines are completely flat.  I don't see any reason to doubt the assumption.

2) Testing for non-linearity

As with mutliple linear regression, we assume that a continuous variable has a linear effect on the response. We can test for that here:

Testing non-linearity

```{r}

martingale_resid <- residuals(c3_fixed, type = "martingale")

plot(data$length_mm, martingale_resid, xlab = "Continuous Covariate",
     ylab = "Martingale Residuals", main = "Continuous Covariate vs. Martingale Residuals")

lines(lowess(data$length_mm, martingale_resid), col = "red")
```

Looks pretty random, seems like the model captured the relationship 

```{r}
martingale_resid <- residuals(c3_fixed, type = "schoenfeld")

data2 <- data %>% 
       filter(treatment == "L") %>% 
       filter(censored == 2) 

plot(data2$length_mm, martingale_resid[,3], xlab = "Continuous Covariate",
     ylab = "Martingale Residuals", main = "Continuous Covariate vs. Martingale Residuals")

lines(lowess(data2$length_mm, martingale_resid[,3]), col = "red")
```

Influential points

```{r}
t<-residuals(c3_fixed,type="deviance")

plot(t)

index <- 1: length(t)
abline(h = 0, lty = 2)  # Add a reference line at y = 0

# Fit and add a loess line
loess_line <- loess(t ~ index, span = 0.6)
lines(index, predict(loess_line), col = "red", lwd = 2)
```

Not many influential points here

# Visualize results

## First save data that is used for graphs

```{r}
saveRDS(data, file = here::here("data","derived","IT_cox_low_results"))
```

```{r}
saveRDS(c3_fixed, file = here::here("data","derived","c3_fixed"))
```

## Region

First, let's compare regions. To do this, construct a new data frame with two rows, one for each value of region.  The other covariates are fixed to their average values (if continuous variables) or to their lowest level (if discrete variables).  For dummy covariate, the average value is the proportion coded 1 in the data set.  

```{r}
region_df <- with(data,
               data.frame(region = c("N", "S"), 
                          length_mm = rep(mean(length_mm, na.rm = TRUE), 2),
                          growout_site = c("CCB", "CCB")
                          )
               )
region_df
```

```{r}
test<-survfit(c3_fixed, newdata = region_df)

p1<-ggsurvplot(test,data=region_df, conf.int=TRUE, ggtheme=theme_bw(),legend.title="Regions",legend.labs = c("N","S"),palette=c("#00BFC4","#F8766D"),title="Low Salinity Challenge",break.x.by=1)
p1
```

Look at data used for the plot

```{r}
p1[["plot"]][["data"]]
```

## Growout_site

```{r}
site_df <- with(data,
               data.frame(region = c("N", "N"), 
                          length_mm = rep(mean(length_mm, na.rm = TRUE), 2),
                          growout_site = c("CCB", "COP")
                          )
               )
site_df
```

```{r}
test<-survfit(c3_fixed, newdata = site_df)

ggsurvplot(test,data=region_df, conf.int=TRUE, ggtheme=theme_bw(),legend.title="Sites",legend.labs = c("CCB","COP"),palette=c("#00BFC4","#F8766D"))
```
